2to3 all

+++ IPTVPlayer/hosts/hostsportdeutschland.py
@@ -67,9 +67,9 @@ class SportDeutschland(CBaseHostClass):
         v = item.get(key, None)
         if None != v:
             try:
-                NumberTypes = (int, long, float, complex)
+                NumberTypes = (int, int, float, complex)
             except NameError:
-                NumberTypes = (int, long, float)
+                NumberTypes = (int, int, float)
                 
             if isinstance(v, NumberTypes):
                 return v
